tools: fix warnings

+++ tools/directfb-csource.c
@@ -457,7 +457,7 @@ static DFBResult load_image (const char            *filename,
 
      data  = malloc (height * pitch);
      if (!data) {
-          fprintf (stderr, "Failed to allocate %lu bytes.\n", height * pitch);
+          fprintf (stderr, "Failed to allocate %lu bytes.\n", (unsigned long)(height * pitch));
           goto cleanup;
      }
 
@@ -502,7 +502,7 @@ static DFBResult load_image (const char            *filename,
 
           dest = malloc (height * d_pitch);
           if (!dest) {
-               fprintf (stderr, "Failed to allocate %lu bytes.\n", height * d_pitch);
+               fprintf (stderr, "Failed to allocate %lu bytes.\n", (unsigned long)(height * d_pitch));
                goto cleanup;
           }
